Commented out unused states. Made a copy of Auto15 called AutoNew in order to successfully sync changes. Commented out Auto15. Attempted to troubleshoot AutoNew.
